move shared CheckDelete behavior to core
