+++ src/magql/sqlalchemy/check_delete.py
@@ -1,16 +1,13 @@
 from __future__ import annotations
 
-import dataclasses
 import typing as t
 
 import sqlalchemy as sa
 import sqlalchemy.orm as sa_orm
 from graphql import GraphQLResolveInfo
 
-from ..core import nodes
-from ..core import scalars
-from ..core.schema import Schema
-from ..core.search import search_result
+from ..core.check_delete import BaseCheckDelete
+from ..core.check_delete import CheckDeleteResult
 from ..core.search import SearchResult
 from ..validators import ValidationError
 
@@ -18,35 +15,12 @@ if t.TYPE_CHECKING:
     from .manager import ModelManager
 
 
-@dataclasses.dataclass()
-class CheckDeleteResult:
-    """The value returned by the :class:`CheckDelete` resolver.
-    :data:`check_delete_result` is the Magql type corresponding to this Python type.
-
-    The items in each list are :class:`.SearchResult`, so that UI behavior can be shared
-    between :class:`CheckDelete` and :class:`.Search`.
-    """
-
-    affected: list[SearchResult] = dataclasses.field(default_factory=list)
-    """Items that will have references to the deleted item removed, such as many-to-many
-    or nullable foreign keys.
-    """
-
-    deleted: list[SearchResult] = dataclasses.field(default_factory=list)
-    """Items that will be deleted along with the deleted item."""
-
-    prevented: list[SearchResult] = dataclasses.field(default_factory=list)
-    """Items that will not be deleted or have references removed prevent the item from
-    being deleted.
-    """
-
-
-class CheckDelete:
+class CheckDelete(BaseCheckDelete):
     """Query field and resolver that shows what would be affected by deleting a row,
     without actually deleting it. Rather than creating a separate query per model, this
     is a generic API using the model name and id.
 
-    The resolver returns :class:`CheckDeleteResult`, which is a collection of
+    The resolver returns :class:`.CheckDeleteResult`, which is a collection of
     :class:`.SearchResult` items.
 
     -   Affected - Items that will have references to the deleted item removed, such as
@@ -64,30 +38,10 @@ class CheckDelete:
     def __init__(
         self, managers: dict[str, ModelManager], field_name: str = "check_delete"
     ) -> None:
+        super().__init__(field_name=field_name)
+
         self.managers = managers
         """Maps model names to managers. These are the models that can be checked."""
-
-        self.field = nodes.Field(
-            check_delete_result,
-            args={
-                "type": nodes.Argument(
-                    scalars.String.non_null, validators=[self._validate_type]
-                ),
-                "id": nodes.Argument(scalars.ID.non_null),
-            },
-            resolve=self,
-        )
-        """The query field.
-
-        .. code-block:: graphql
-
-            type Query {
-                check_delete(type: String!, id: ID!): CheckDeleteResult!
-            }
-        """
-
-        self.field_name = field_name
-        """The name to use for this field in the top-level query object."""
 
     def _validate_type(self, info: GraphQLResolveInfo, value: str, data: t.Any) -> None:
         if value not in self.managers:
@@ -148,29 +102,3 @@ class CheckDelete:
                 )
 
         return result
-
-    def register(self, schema: Schema) -> None:
-        schema.query.fields[self.field_name] = self.field
-
-
-check_delete_result = nodes.Object(
-    "CheckDeleteResult",
-    fields={
-        "affected": search_result.non_null.list.non_null,
-        "deleted": search_result.non_null.list.non_null,
-        "prevented": search_result.non_null.list.non_null,
-    },
-)
-"""The result type for the :class:`CheckDelete` query. :class:`CheckDeleteResult` is the
-Python type corresponding to this Magql type.
-
-The items in each list are :class:`.search_result`.
-
-.. code-block:: graphql
-
-    type CheckDeleteResult {
-        affected: [SearchResult!]!
-        deleted: [SearchResult!]!
-        prevented: [SearchResult!]!
-    }
-"""
